+++ packages/react-native/src/icons/OpenRouter/components/Color.tsx
@@ -1,0 +1,19 @@
+import React, { memo } from 'react';
+import { Path, Svg } from 'react-native-svg';
+
+import type { RNIconProps } from '@/features';
+
+const Icon = memo<RNIconProps>(({ size = 24, style, ...rest }) => {
+  return (
+    <Svg height={size} style={style} viewBox="0 0 24 24" width={size} {...rest}>
+      <Path
+        d="M18.654 3.87a5.087 5.087 0 110 10.174L23.7 19.09c.64.641.187 1.737-.72 1.737H8.48a8.479 8.479 0 010-16.958h10.175zM8.479 7.26a5.087 5.087 0 100 10.176 5.087 5.087 0 000-10.175z"
+        fill="#C8FF00"
+      />
+    </Svg>
+  );
+});
+
+Icon.displayName = 'OpenRouterColor';
+
+export default Icon;