+++ packages/react-native/src/icons/Pi/components/Mono.tsx
@@ -1,0 +1,29 @@
+import React, { memo } from 'react';
+import { Path, Svg } from 'react-native-svg';
+
+import type { RNIconProps } from '@/features';
+
+const Icon = memo<RNIconProps>(({ size = 24, style, color = '#000000', ...rest }) => {
+  return (
+    <Svg
+      color={color}
+      fillRule="evenodd"
+      height={size}
+      style={style}
+      viewBox="0 0 24 24"
+      width={size}
+      {...rest}
+    >
+      <Path
+        clipRule="evenodd"
+        d="M1 1h16.5v11H12v5.5H6.5V23H1V1zm5.5 5.5V12H12V6.5H6.5z"
+        fill={color}
+      />
+      <Path d="M17.5 12H23v11h-5.5V12z" fill={color} />
+    </Svg>
+  );
+});
+
+Icon.displayName = 'PiMono';
+
+export default Icon;