+++ src/OpenRouter/style.ts
@@ -1,9 +1,9 @@
 export const TITLE = 'OpenRouter';
 export const COMBINE_TEXT_MULTIPLE = 1;
-export const COMBINE_SPACE_MULTIPLE = 0.4;
-export const COLOR_PRIMARY = '#6566F1';
+export const COMBINE_SPACE_MULTIPLE = 0.3;
+export const COLOR_PRIMARY = '#C8FF00';
 
 // Avatar constants
-export const AVATAR_BACKGROUND = COLOR_PRIMARY;
-export const AVATAR_COLOR = '#fff';
+export const AVATAR_BACKGROUND = '#000';
+export const AVATAR_COLOR = COLOR_PRIMARY;
 export const AVATAR_ICON_MULTIPLE = 0.75;