+++ CHANGELOG.md
@@ -2,6 +2,111 @@
 
 # Changelog
 
+## [Version 5.13.0](https://github.com/lobehub/lobe-icons/compare/v5.12.0...v5.13.0)
+
+<sup>Released on **2026-07-14**</sup>
+
+#### ✨ Features
+
+- **auto**: Auto build static icons.
+- **misc**: Add Anspire, Bocha, Brave, Browserless, Kagi, SearXNG.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's improved
+
+- **auto**: Auto build static icons ([ef22550](https://github.com/lobehub/lobe-icons/commit/ef22550))
+- **misc**: Add Anspire, Bocha, Brave, Browserless, Kagi, SearXNG ([ba5e7e5](https://github.com/lobehub/lobe-icons/commit/ba5e7e5))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version 5.12.0](https://github.com/lobehub/lobe-icons/compare/v5.11.0...v5.12.0)
+
+<sup>Released on **2026-07-14**</sup>
+
+#### ✨ Features
+
+- **misc**: Add Poolside, Firecrawl, AgnesAI, Pi, HappyHorse, Celesto.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's improved
+
+- **misc**: Add Poolside, Firecrawl, AgnesAI, Pi, HappyHorse, Celesto, closes [#360](https://github.com/lobehub/lobe-icons/issues/360) [#323](https://github.com/lobehub/lobe-icons/issues/323) [#322](https://github.com/lobehub/lobe-icons/issues/322) [#349](https://github.com/lobehub/lobe-icons/issues/349) [#348](https://github.com/lobehub/lobe-icons/issues/348) ([6274a43](https://github.com/lobehub/lobe-icons/commit/6274a43))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version 5.11.0](https://github.com/lobehub/lobe-icons/compare/v5.10.1...v5.11.0)
+
+<sup>Released on **2026-07-11**</sup>
+
+#### ✨ Features
+
+- **misc**: Add Grok icon mapping for supergrok provider, add Grok icon mapping for supergrok provider in react-native package.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's improved
+
+- **misc**: Add Grok icon mapping for supergrok provider ([0a6f356](https://github.com/lobehub/lobe-icons/commit/0a6f356))
+- **misc**: Add Grok icon mapping for supergrok provider in react-native package ([793d97e](https://github.com/lobehub/lobe-icons/commit/793d97e))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+### [Version 5.10.1](https://github.com/lobehub/lobe-icons/compare/v5.10.0...v5.10.1)
+
+<sup>Released on **2026-07-01**</sup>
+
+#### 🐛 Bug Fixes
+
+- **misc**: Bound NanoBanana regex to avoid matching unrelated Gemini image models, match Nano Banana 2 Lite model id to NanoBanana icon, sync NanoBanana regex to React Native model mapping.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's fixed
+
+- **misc**: Bound NanoBanana regex to avoid matching unrelated Gemini image models, closes [#347](https://github.com/lobehub/lobe-icons/issues/347) ([547c5c8](https://github.com/lobehub/lobe-icons/commit/547c5c8))
+- **misc**: Match Nano Banana 2 Lite model id to NanoBanana icon ([38edfd7](https://github.com/lobehub/lobe-icons/commit/38edfd7))
+- **misc**: Sync NanoBanana regex to React Native model mapping, closes [#347](https://github.com/lobehub/lobe-icons/issues/347) ([78766a1](https://github.com/lobehub/lobe-icons/commit/78766a1))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
 ## [Version 5.10.0](https://github.com/lobehub/lobe-icons/compare/v5.9.0...v5.10.0)
 
 <sup>Released on **2026-05-28**</sup>