+++ packages/react-native/CHANGELOG.md
@@ -2,6 +2,107 @@
 
 # Changelog
 
+## [Version 2.11.0](https://github.com/lobehub/lobe-icons/compare/@lobehub/icons-rn@2.10.0...@lobehub/icons-rn@2.11.0)
+
+<sup>Released on **2026-07-14**</sup>
+
+#### ✨ Features
+
+- **misc**: Add Anspire, Bocha, Brave, Browserless, Kagi, SearXNG.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's improved
+
+- **misc**: Add Anspire, Bocha, Brave, Browserless, Kagi, SearXNG ([ba5e7e5](https://github.com/lobehub/lobe-icons/commit/ba5e7e5))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version 2.10.0](https://github.com/lobehub/lobe-icons/compare/@lobehub/icons-rn@2.9.0...@lobehub/icons-rn@2.10.0)
+
+<sup>Released on **2026-07-14**</sup>
+
+#### ✨ Features
+
+- **misc**: Add Poolside, Firecrawl, AgnesAI, Pi, HappyHorse, Celesto.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's improved
+
+- **misc**: Add Poolside, Firecrawl, AgnesAI, Pi, HappyHorse, Celesto, closes [#360](https://github.com/lobehub/lobe-icons/issues/360) [#323](https://github.com/lobehub/lobe-icons/issues/323) [#322](https://github.com/lobehub/lobe-icons/issues/322) [#349](https://github.com/lobehub/lobe-icons/issues/349) [#348](https://github.com/lobehub/lobe-icons/issues/348) ([6274a43](https://github.com/lobehub/lobe-icons/commit/6274a43))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version 2.9.0](https://github.com/lobehub/lobe-icons/compare/@lobehub/icons-rn@2.8.1...@lobehub/icons-rn@2.9.0)
+
+<sup>Released on **2026-07-11**</sup>
+
+#### ✨ Features
+
+- **misc**: Add Grok icon mapping for supergrok provider in react-native package.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's improved
+
+- **misc**: Add Grok icon mapping for supergrok provider in react-native package ([793d97e](https://github.com/lobehub/lobe-icons/commit/793d97e))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+### [Version 2.8.1](https://github.com/lobehub/lobe-icons/compare/@lobehub/icons-rn@2.8.0...@lobehub/icons-rn@2.8.1)
+
+<sup>Released on **2026-07-01**</sup>
+
+#### 🐛 Bug Fixes
+
+- **misc**: Bound NanoBanana regex to avoid matching unrelated Gemini image models, sync NanoBanana regex to React Native model mapping.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's fixed
+
+- **misc**: Bound NanoBanana regex to avoid matching unrelated Gemini image models, closes [#347](https://github.com/lobehub/lobe-icons/issues/347) ([547c5c8](https://github.com/lobehub/lobe-icons/commit/547c5c8))
+- **misc**: Sync NanoBanana regex to React Native model mapping, closes [#347](https://github.com/lobehub/lobe-icons/issues/347) ([78766a1](https://github.com/lobehub/lobe-icons/commit/78766a1))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
 ## [Version 2.8.0](https://github.com/lobehub/lobe-icons/compare/@lobehub/icons-rn@2.7.1...@lobehub/icons-rn@2.8.0)
 
 <sup>Released on **2026-05-28**</sup>